Clean up and add some Javadoc.
