+++ gcd/src/main/java/org/spine3/server/storage/datastore/DsRecordStorage.java
@@ -253,7 +253,7 @@ public class DsRecordStorage<I> extends RecordStorage<I> {
     private Collection<Filter> buildColumnFilters(
             Iterable<CompositeQueryParameter> compositeParameters) {
         final Collection<CompositeQueryParameter> params = newArrayList(compositeParameters);
-        final Collection<Filter> predicate = DsQueries.fromParams(params, columnHandler);
+        final Collection<Filter> predicate = DsFilters.fromParams(params, columnHandler);
         return predicate;
     }
